third_party/WebKit/LayoutTests/harness-tests/mojo-helpers.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script>
'use strict';

// Verify that the mojo_test helper functions properly and Mojo bindings
// are available.
mojo_test(mojo => {
  assert_true(mojo instanceof Object);
  assert_true(mojo.core instanceof Object);
  assert_true(mojo.router instanceof Object);
  assert_true(mojo.serviceRegistry instanceof Object);
}, 'Mojo system APIs should be available to layout tests.');
</script>
